+++ syslog/main.py
@@ -28,19 +28,18 @@ SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
 CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY,
 OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
 OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-"""
 
-"""Syslog Plugin."""
-
+Syslog Plugin."""
 
 import logging
 import logging.handlers
 import threading
 import socket
 import json
+import traceback
+import time
 from typing import List
 from jsonpath import jsonpath
-
 from netskope.common.utils import AlertsHelper
 from netskope.integrations.cls.plugin_base import (
     PluginBase,
@@ -48,29 +47,60 @@ from netskope.integrations.cls.plugin_base import (
     PushResult,
 )
 from .utils.syslog_constants import (
-    SYSLOG_FORMATS,
     SYSLOG_PROTOCOLS,
-    PLUGIN_NAME,
+    PLATFORM_NAME,
+    MODULE_NAME,
+    PLUGIN_VERSION,
 )
-from .utils.syslog_validator import (
-    SyslogValidator,
-)
-from .utils.syslog_helper import (
-    get_syslog_mappings,
-)
+from .utils.syslog_validator import SyslogValidator
+from .utils.syslog_helper import get_syslog_mappings
 from .utils.syslog_exceptions import (
     MappingValidationError,
     EmptyExtensionError,
     FieldNotFoundError,
+    SyslogPluginError,
 )
-from .utils.syslog_cef_generator import (
-    CEFGenerator,
-)
+from .utils.syslog_cef_generator import CEFGenerator
 from .utils.syslog_ssl import SSLSysLogHandler
 
 
 class SyslogPlugin(PluginBase):
     """The Syslog plugin implementation class."""
+
+    def __init__(
+        self,
+        name,
+        *args,
+        **kwargs,
+    ):
+        """Initialize SyslogPlugin class."""
+        super().__init__(
+            name,
+            *args,
+            **kwargs,
+        )
+        self.plugin_name, self.plugin_version = self._get_plugin_info()
+        self.log_prefix = f"{MODULE_NAME} {self.plugin_name} [{name}]"
+
+    def _get_plugin_info(self) -> tuple:
+        """Get plugin name and version from manifest.
+        Returns:
+            tuple: Tuple of plugin's name and version fetched from manifest.
+        """
+        try:
+            manifest_json = SyslogPlugin.metadata
+            plugin_name = manifest_json.get("name", PLATFORM_NAME)
+            plugin_version = manifest_json.get("version", PLUGIN_VERSION)
+            return (plugin_name, plugin_version)
+        except Exception as exp:
+            self.logger.error(
+                message=(
+                    f"{self.log_prefix}: Error occurred while"
+                    f" getting plugin details. Error: {exp}"
+                ),
+                details=str(traceback.format_exc()),
+            )
+        return (PLATFORM_NAME, PLUGIN_VERSION)
 
     def get_mapping_value_from_json_path(self, data, json_path):
         """To Fetch the value from given JSON object using given JSON path.
@@ -95,9 +125,9 @@ class SyslogPlugin(PluginBase):
             fetched value.
         """
         return (
-            data[field]
+            (data[field], True)
             if data[field] or isinstance(data[field], int)
-            else "null"
+            else ("null", False)
         )
 
     def get_subtype_mapping(self, mappings, subtype):
@@ -139,25 +169,30 @@ class SyslogPlugin(PluginBase):
             mapping_variables = {"$tenant_name": tenant.name}
 
         missing_fields = []
+        mapped_field_flag = False
         # Iterate over mapped headers
         for cef_header, header_mapping in header_mappings.items():
             try:
-                headers[cef_header] = self.get_field_value_from_data(
+                (
+                    headers[cef_header],
+                    mapped_field,
+                ) = self.get_field_value_from_data(
                     header_mapping, data, data_type, subtype, False
                 )
+
+                if mapped_field:
+                    mapped_field_flag = mapped_field
 
                 # Handle variable mappings
                 if (
                     isinstance(headers[cef_header], str)
                     and headers[cef_header].lower() in mapping_variables
                 ):
-                    headers[cef_header] = mapping_variables[
-                        headers[cef_header].lower()
-                    ]
+                    headers[cef_header] = mapping_variables[headers[cef_header].lower()]
             except FieldNotFoundError as err:
                 missing_fields.append(str(err))
 
-        return headers
+        return headers, mapped_field_flag
 
     def get_extensions(self, extension_mappings, data, data_type, subtype):
         """Fetch extensions from given mappings.
@@ -173,21 +208,28 @@ class SyslogPlugin(PluginBase):
         """
         extension = {}
         missing_fields = []
+        mapped_field_flag = False
 
         # Iterate over mapped extensions
         for cef_extension, extension_mapping in extension_mappings.items():
             try:
-                extension[cef_extension] = self.get_field_value_from_data(
+                (
+                    extension[cef_extension],
+                    mapped_field,
+                ) = self.get_field_value_from_data(
                     extension_mapping,
                     data,
                     data_type,
                     subtype,
                     is_json_path="is_json_path" in extension_mapping,
                 )
+
+                if mapped_field:
+                    mapped_field_flag = mapped_field
             except FieldNotFoundError as err:
                 missing_fields.append(str(err))
 
-        return extension
+        return extension, mapped_field_flag
 
     def get_field_value_from_data(
         self, extension_mapping, data, data_type, subtype, is_json_path=False
@@ -221,10 +263,10 @@ class SyslogPlugin(PluginBase):
            NP    |     NP     |        NP      |           - (Not possible)
         -----------------------------------------------------------------------
         """
-        if (
-            "mapping_field" in extension_mapping
-            and extension_mapping["mapping_field"]
-        ):
+        # mapped_field will be returned as true only if the value returned is\
+        # using the mapping_field and not default_value
+        mapped_field = False
+        if "mapping_field" in extension_mapping and extension_mapping["mapping_field"]:
             if is_json_path:
                 # If mapping field specified by JSON path is present in data,\
                 #  map that field, else skip by raising
@@ -233,24 +275,23 @@ class SyslogPlugin(PluginBase):
                     data, extension_mapping["mapping_field"]
                 )
                 if value:
-                    return ",".join([str(val) for val in value])
+                    mapped_field = True
+                    return ",".join([str(val) for val in value]), mapped_field
                 else:
-                    raise FieldNotFoundError(
-                        extension_mapping["mapping_field"]
-                    )
+                    raise FieldNotFoundError(extension_mapping["mapping_field"])
             else:
                 # If mapping is present in data, map that field, \
                 # else skip by raising exception
-                if (
-                    extension_mapping["mapping_field"] in data
-                ):  # case #1 and case #4
+                if extension_mapping["mapping_field"] in data:  # case #1 and case #4
                     if (
                         extension_mapping.get("transformation") == "Time Stamp"
                         and data[extension_mapping["mapping_field"]]
                     ):
                         try:
-                            return int(
-                                data[extension_mapping["mapping_field"]]
+                            mapped_field = True
+                            return (
+                                int(data[extension_mapping["mapping_field"]]),
+                                mapped_field,
                             )
                         except Exception:
                             pass
@@ -260,17 +301,15 @@ class SyslogPlugin(PluginBase):
                 elif "default_value" in extension_mapping:
                     # If mapped value is not found in response and default is \
                     # mapped, map the default value (case #2)
-                    return extension_mapping["default_value"]
+                    return extension_mapping["default_value"], mapped_field
                 else:  # case #6
-                    raise FieldNotFoundError(
-                        extension_mapping["mapping_field"]
-                    )
+                    raise FieldNotFoundError(extension_mapping["mapping_field"])
         else:
             # If mapping is not present, 'default_value' must be there\
             #  because of validation (case #3 and case #5)
-            return extension_mapping["default_value"]
+            return extension_mapping["default_value"], mapped_field
 
-    def map_json_data(self, mappings, data, data_type, subtype):
+    def map_json_data(self, mappings, data):
         """Filter the raw data and returns the filtered data.
 
         :param mappings: List of fields to be pushed
@@ -279,7 +318,7 @@ class SyslogPlugin(PluginBase):
         :return: Mapped data based on fields given in mapping file
         """
 
-        if mappings == []:
+        if mappings == [] or not data:
             return data
 
         mapped_dict = {}
@@ -292,53 +331,109 @@ class SyslogPlugin(PluginBase):
     def transform(self, raw_data, data_type, subtype) -> List:
         """To Transform the raw netskope JSON data into target platform \
             supported data formats."""
-
+        count = 0
+        log_source_identifier = self.configuration.get(
+            "log_source_identifier", "netskopece"
+        )
         if not self.configuration.get("transformData", True):
-            if data_type not in ["alerts", "events"]:
-                return raw_data
-
             try:
                 delimiter, cef_version, syslog_mappings = get_syslog_mappings(
                     self.mappings, "json", self.name
                 )
             except KeyError as err:
-                self.logger.error(
-                    f"{PLUGIN_NAME}[{self.name}]: Error in syslog mapping file."
-                    f"Error: {err}"
+                error_msg = (
+                    f"[{data_type}]:[{subtype}] "
+                    "An error occurred while fetching the mappings."
                 )
-                raise
+                self.logger.error(
+                    message=(
+                        f"{self.log_prefix}: {error_msg} Error: {err}"
+                    ),
+                    details=str(traceback.format_exc()),
+                )
+                raise SyslogPluginError(error_msg)
             except MappingValidationError as err:
-                self.logger.error(
-                    f"{PLUGIN_NAME}[{self.name}]: Validation error occurred. {err}"
+                error_msg = (
+                    f"[{data_type}]:[{subtype}] "
+                    "An error occurred while validating the mapping file."
                 )
-                raise
+                self.logger.error(
+                    message=(
+                        f"{self.log_prefix}: {error_msg} {err}"
+                    ),
+                    details=str(traceback.format_exc()),
+                )
+                raise SyslogPluginError(error_msg)
             except Exception as err:
-                self.logger.error(
-                    f"{PLUGIN_NAME}[{self.name}]: An error occurred while mapping data\
-                          using given json mappings. Error: {err}"
+                error_msg = (
+                    f"[{data_type}]:[{subtype}] "
+                    "An error occurred while mapping "
+                    "data using given json mappings."
                 )
-                raise
+                self.logger.error(
+                    message=(
+                        f"{self.log_prefix}: {error_msg} Error: {err}"
+                    ),
+                    details=str(traceback.format_exc()),
+                )
+                raise SyslogPluginError(error_msg)
 
             try:
                 subtype_mapping = self.get_subtype_mapping(
                     syslog_mappings["json"][data_type], subtype
                 )
-            except Exception:
-                self.logger.error(
-                    f"{PLUGIN_NAME}[{self.name}]: Error occurred while retrieving mappings \
-                        for datatype: {data_type} (subtype: {subtype}) \
-                            Transformation will be skipped."
-                )
+                if subtype_mapping == []:
+                    transformed_data = []
+                    for data in raw_data:
+                        if data:                            
+                            result = "{} {} {}".format(
+                                time.strftime(
+                                    "%b %d %H:%M:%S", time.localtime(time.time())
+                                ),
+                                log_source_identifier,
+                                json.dumps(data),
+                            )
+                            transformed_data.append(result)
+                        else:
+                            count += 1
+                    return transformed_data
+            except SyslogPluginError:
                 raise
+            except Exception as err:
+                error_msg = (
+                    f"[{data_type}]:[{subtype}] "
+                    "Error occurred while retrieving "
+                    f"mappings."
+                )
+                self.logger.error(
+                    message=(
+                        f"{self.log_prefix}: {error_msg} Error: {err}"
+                    ),
+                    details=str(traceback.format_exc()),
+                )
+                raise SyslogPluginError(error_msg)
 
             transformed_data = []
-
             for data in raw_data:
-                mapped_dict = self.map_json_data(
-                    subtype_mapping, data, data_type, subtype
-                )
+                mapped_dict = self.map_json_data(subtype_mapping, data)
                 if mapped_dict:
-                    transformed_data.append(mapped_dict)
+                    result = "{} {} {}".format(
+                        time.strftime("%b %d %H:%M:%S", time.localtime(time.time())),
+                        log_source_identifier,
+                        json.dumps(mapped_dict),
+                    )
+                    transformed_data.append(result)
+                else:
+                    count += 1
+
+            if count > 0:
+                self.logger.info(
+                    "{}: Plugin couldn't process {} records because they "
+                    "either had no data or contained invalid/missing "
+                    "fields according to the configured JSON mapping. "
+                    "Therefore, the transformation and ingestion for those "
+                    "records were skipped.".format(self.log_prefix, count)
+                )
 
             return transformed_data
         else:
@@ -347,107 +442,174 @@ class SyslogPlugin(PluginBase):
                     self.mappings, data_type, self.name
                 )
             except KeyError as err:
-                self.logger.error(
-                    f"{PLUGIN_NAME}[{self.name}]: Error in syslog mapping file."
-                    f"Error: {err}"
+                error_msg = (
+                    f"[{data_type}]:[{subtype}] "
+                    "An error occurred while "
+                    "fetching the mappings."
                 )
-                raise
+                self.logger.error(
+                    message=(
+                        f"{self.log_prefix}: {error_msg} "
+                        f"Error: {err}"
+                    ),
+                    details=str(traceback.format_exc()),
+                )
+                raise SyslogPluginError(error_msg)
             except MappingValidationError as err:
-                self.logger.error(
-                    f"{PLUGIN_NAME}[{self.name}]: Validation error occurred. {err}"
+                error_msg = (
+                    f"[{data_type}]:[{subtype}] "
+                    "An error occurred while validating the mapping file."
                 )
-                raise
+                self.logger.error(
+                    message=(
+                        f"{self.log_prefix}: {error_msg} {err}"
+                    ),
+                    details=str(traceback.format_exc()),
+                )
+                raise SyslogPluginError(error_msg)
             except Exception as err:
-                self.logger.error(
-                    f"{PLUGIN_NAME}[{self.name}]: An error occurred while mapping data \
-                        using given json mappings. Error: {err}"
+                error_msg = (
+                    f"[{data_type}]:[{subtype}] "
+                    "An error occurred while mapping "
+                    "data using given json mappings."
                 )
-                raise
+                self.logger.error(
+                    message=(
+                        f"{self.log_prefix}: {error_msg} Error: {err}"
+                    ),
+                    details=str(traceback.format_exc()),
+                )
+                raise SyslogPluginError(error_msg)
 
             cef_generator = CEFGenerator(
-                self.mappings, delimiter, cef_version, self.logger, self.name
+                self.mappings,
+                delimiter,
+                cef_version,
+                self.logger,
+                self.log_prefix,
             )
+
+            # First retrieve the mapping of subtype being transformed
+            try:
+                subtype_mapping = self.get_subtype_mapping(
+                    syslog_mappings[data_type], subtype
+                )
+            except KeyError as err:
+                error_msg = (
+                    f"[{data_type}]:[{subtype}] Unable to find the "
+                    "mappings in the mapping file."
+                )
+                self.logger.info(
+                    f"{self.log_prefix}: {error_msg}"
+                    f"Error: {err}"
+                )
+                raise SyslogPluginError(error_msg)
+            except Exception as err:
+                error_msg = (
+                    f"[{data_type}]:[{subtype}] Error occurred while retrieving "
+                    f"mappings for subtype {subtype}."
+                )
+                self.logger.error(
+                    message=(
+                        f"{self.log_prefix}: {error_msg} Error: {err}"
+                    ),
+                    details=str(traceback.format_exc()),
+                )
+                raise SyslogPluginError(error_msg)
 
             transformed_data = []
             for data in raw_data:
-                # First retrieve the mapping of subtype being transformed
-                try:
-                    subtype_mapping = self.get_subtype_mapping(
-                        syslog_mappings[data_type], subtype
-                    )
-                except Exception:
-                    self.logger.error(
-                        f"{PLUGIN_NAME}[{self.name}]: Error occurred while retrieving \
-                            mappings for subtype {subtype}. \
-                            Transformation of current record will be skipped."
-                    )
+                if not data:
+                    count += 1
                     continue
 
                 # Generating the CEF header
                 try:
-                    header = self.get_headers(
+                    header, mapped_flag_header = self.get_headers(
                         subtype_mapping["header"], data, data_type, subtype
                     )
                 except Exception as err:
                     self.logger.error(
-                        f"{PLUGIN_NAME}[{self.name}]([{data_type}][{subtype}]): Error \
-                            occurred while creating CEF header: {err}. \
-                            Transformation of current record will be skipped."
+                        message=(
+                            f"{self.log_prefix}: [{data_type}][{subtype}] "
+                            f"Error occurred while creating CEF header: {err}."
+                            " Transformation of current record will be skipped."
+                        ),
+                        details=str(traceback.format_exc()),
                     )
+                    count += 1
                     continue
 
                 try:
-                    extension = self.get_extensions(
+                    extension, mapped_flag_extension = self.get_extensions(
                         subtype_mapping["extension"], data, data_type, subtype
                     )
                 except Exception as err:
                     self.logger.error(
-                        f"{PLUGIN_NAME}[{self.name}]([{data_type}][{subtype}]): Error \
-                        occurred while creating CEF extension: {err}. \
-                        Transformation of the current record will be skipped"
+                        message=(
+                            f"{self.log_prefix}: [{data_type}][{subtype}] Error "
+                            f"occurred while creating CEF extension: {err}. "
+                            "Transformation of the current record will be skipped."
+                        ),
+                        details=str(traceback.format_exc()),
                     )
+                    count += 1
                     continue
 
                 try:
+                    if not (mapped_flag_header or mapped_flag_extension):
+                        count += 1
+                        continue
                     cef_generated_event = cef_generator.get_cef_event(
                         data,
                         header,
                         extension,
                         data_type,
                         subtype,
-                        self.configuration.get(
-                            "log_source_identifier", "netskopece"
-                        ),
+                        log_source_identifier,
                     )
                     if cef_generated_event:
                         transformed_data.append(cef_generated_event)
                 except EmptyExtensionError:
                     self.logger.error(
-                        f"{PLUGIN_NAME}[{self.name}]([{data_type}][{subtype}]): Got empty \
-                            extension during transformation."
-                        "Transformation of current record will be skipped"
+                        message=(
+                            f"{self.log_prefix}: [{data_type}][{subtype}] Got"
+                            " empty extension during transformation. "
+                            "Transformation of current record will be skipped."
+                        ),
+                        details=str(traceback.format_exc()),
                     )
+                    count += 1
                 except Exception as err:
                     self.logger.error(
-                        f"{PLUGIN_NAME}[{self.name}]([{data_type}][{subtype}]): An error \
-                            occurred during transformation. Error: {err}"
+                        message=(
+                            f"{self.log_prefix}: [{data_type}][{subtype}] An "
+                            f"error occurred during transformation. Error: {err}"
+                        ),
+                        details=str(traceback.format_exc()),
                     )
+                    count += 1
+            if count > 0:
+                self.logger.debug(
+                    "{}: Plugin couldn't process {} records because they "
+                    "either had no data or contained invalid/missing "
+                    "fields according to the configured mapping. "
+                    "Therefore, the transformation and ingestion for those "
+                    "records were skipped.".format(self.log_prefix, count)
+                )
 
             return transformed_data
 
     def init_handler(self, configuration):
         """Initialize unique Syslog handler per thread\
               based on configured protocol."""
-        syslogger = logging.getLogger(
-            "SYSLOG_LOGGER_{}".format(threading.get_ident())
-        )
+        syslogger = logging.getLogger("SYSLOG_LOGGER_{}".format(threading.get_ident()))
         syslogger.setLevel(logging.INFO)
         syslogger.handlers = []
         syslogger.propagate = False
 
         if configuration["syslog_protocol"] == "TLS":
             tls_handler = SSLSysLogHandler(
-                configuration.get("transformData", True),
                 configuration["syslog_protocol"],
                 address=(
                     configuration["syslog_server"].strip(),
@@ -463,7 +625,6 @@ class SyslogPlugin(PluginBase):
 
             # Create a syslog handler with given configuration parameters
             handler = SSLSysLogHandler(
-                configuration.get("transformData", True),
                 configuration["syslog_protocol"],
                 address=(
                     configuration["syslog_server"].strip(),
@@ -485,54 +646,144 @@ class SyslogPlugin(PluginBase):
             syslogger.addHandler(handler)
 
         return syslogger
+    
+    def log_success_msg(
+            self,
+            data_type,
+            subtype,
+            successful_log_push_counter,
+            skipped_logs
+    ):
+        """Log the success message."""
+        log_msg = ""
+        if successful_log_push_counter > 0:
+            log_msg = (
+                f"[{data_type}] [{subtype}] Successfully "
+                f"ingested {successful_log_push_counter} log(s)"
+                f" to {self.plugin_name} server."
+            )
+        if log_msg and skipped_logs:
+            log_msg += (
+                " Received empty transformed data for "
+                f"{skipped_logs} log(s) hence ingestion of those log(s) "
+                "will be skipped."
+            )
+        if log_msg:
+            self.logger.info(
+                f"{self.log_prefix}: {log_msg}"
+            )
+        return log_msg
 
     def push(self, transformed_data, data_type, subtype) -> PushResult:
         """Push the transformed_data to the 3rd party platform."""
+        self.logger.debug(
+            f"{self.log_prefix}: "
+            f"Initializing the sharing of {len(transformed_data)} "
+            f"[{data_type}]:[{subtype}] logs "
+            f"to the {PLATFORM_NAME} server."
+        )
+        successful_log_push_counter, skipped_logs = 0, 0
         try:
             syslogger = self.init_handler(self.configuration)
-        except Exception as err:
+        except (TimeoutError, ConnectionRefusedError, Exception) as err:
+            error_msg = {
+                TimeoutError: (
+                    "Timeout error occurred during initializing "
+                    "connection with Syslog server."
+                ),
+                ConnectionRefusedError: (
+                    "Connection refused during initializing "
+                    "connection with Syslog server."
+                ),
+            }.get(type(err), "Error occurred during initializing connection.")
             self.logger.error(
-                f"{PLUGIN_NAME}[{self.name}]: Error occurred during \
-                    initializing connection. Error: {err}"
+                message=f"{self.log_prefix}: {error_msg} {err}.",
+                details=str(traceback.format_exc()),
             )
-            raise
+            raise SyslogPluginError(error_msg)
 
         # Log the transformed data to given syslog server
-        for data in transformed_data:
-            try:
+        try:
+            for data in transformed_data:
                 if data:
                     syslogger.info(
                         json.dumps(data) if isinstance(data, dict) else data
                     )
                     syslogger.handlers[0].flush()
+                    successful_log_push_counter += 1
+                else:
+                    skipped_logs += 1
+            
+            log_msg = self.log_success_msg(
+                data_type,
+                subtype,
+                successful_log_push_counter,
+                skipped_logs
+            )
+            return PushResult(
+                success=True,
+                message=log_msg,
+            )
 
-            except Exception as err:
-                self.logger.error(
-                    f"{PLUGIN_NAME}[{self.name}]: Error occurred during data ingestion."
-                    f"Error: {err}. Record will be skipped"
-                )
+        except Exception as err:
+            failed_logs = (
+                len(transformed_data) - successful_log_push_counter
+            )
+            _ = self.log_success_msg(
+                data_type,
+                subtype,
+                successful_log_push_counter,
+                skipped_logs
+            )
+            # Remove the data already been ingested
+            del transformed_data[:successful_log_push_counter]
+            error_msg = (
+                "Error occurred while ingesting "
+                f"data to Syslog server. {failed_logs} log(s) "
+                "failed to be ingested."
+            )
+            self.logger.error(
+                message=(
+                    f"{self.log_prefix}: {error_msg}"
+                    f" Error: {err}"
+                ),
+                details=str(traceback.format_exc()),
+            )
+            raise SyslogPluginError(error_msg)
 
         # Clean up
-        try:
-            syslogger.handlers[0].close()
-            del syslogger.handlers[:]
-            del syslogger
-        except Exception as err:
-            self.logger.error(
-                f"{PLUGIN_NAME}[{self.name}]: Error occurred during Clean up. Error: {err}"
-            )
+        finally:
+            try:
+                syslogger.handlers[0].close()
+                del syslogger.handlers[:]
+                del syslogger
+            except Exception as err:
+                error_msg = (
+                    "Error occurred while cleaning up "
+                    "Syslog handler."
+                )
+                self.logger.debug(
+                    f"{self.log_prefix}: {error_msg} "
+                    f"Error: {err}"
+                )
 
     def test_server_connectivity(self, configuration):
         """Tests whether the configured syslog server is reachable or not."""
         try:
             syslogger = self.init_handler(configuration)
         except Exception as err:
-            self.logger.error(
-                f"{PLUGIN_NAME}[{self.name}]: Error occurred while establishing connection \
-                    with syslog server. Make sure "
+            error_msg = (
+                "Error occurred while establishing "
+                "connection with syslog server. Make sure "
                 "you have provided correct syslog server and port."
             )
-            raise err
+            self.logger.error(
+                message=(
+                    f"{self.log_prefix}: {error_msg} Error: {err}"
+                ),
+                details=str(traceback.format_exc()),
+            )
+            raise SyslogPluginError(error_msg)
         else:
             # Clean up for further use
             syslogger.handlers[0].flush()
@@ -542,178 +793,137 @@ class SyslogPlugin(PluginBase):
 
     def validate(self, configuration: dict) -> ValidationResult:
         """Validate the configuration parameters dict."""
-        syslog_validator = SyslogValidator(self.logger, self.name)
-        if (
-            "syslog_server" not in configuration
-            or not configuration["syslog_server"].strip()
-        ):
-            self.logger.error(
-                f"{PLUGIN_NAME}[{self.name}]: Validation error occurred. Error: "
-                "Syslog Server IP/FQDN is a required field in \
-                    the configuration parameters."
-            )
+        syslog_validator = SyslogValidator(self.logger, self.log_prefix)
+        validation_err_msg = f"{self.log_prefix}: Validation error occurred."
+        syslog_server = configuration.get("syslog_server", "").strip()
+        if not syslog_server:
+            err_msg = "Syslog Server is a required configuration parameter."
+            self.logger.error(f"{validation_err_msg} {err_msg}")
             return ValidationResult(
-                success=False, message="Syslog Server is a required field."
+                success=False,
+                message=err_msg,
             )
-        elif type(configuration["syslog_server"].strip()) != str:
-            self.logger.error(
-                f"{PLUGIN_NAME}[{self.name}]: Validation error occurred. Error: "
-                "Invalid Syslog Server IP/FQDN found \
-                    in the configuration parameters."
-            )
+        elif not isinstance(syslog_server, str):
+            err_msg = "Invalid Syslog Server provided in configuration parameters."
+            self.logger.error(f"{validation_err_msg} {err_msg}")
             return ValidationResult(
-                success=False, message="Invalid Syslog Server provided."
+                success=False,
+                message=err_msg,
             )
-        if (
-            "syslog_format" not in configuration
-            or not configuration["syslog_format"].strip()
-        ):
-            self.logger.error(
-                f"{PLUGIN_NAME}[{self.name}]: Validation error occurred. Error: "
-                "Syslog Format is a required field in \
-                    the configuration parameters."
-            )
+
+        syslog_protocol = configuration.get("syslog_protocol", "").strip()
+        if not syslog_protocol:
+            err_msg = "Syslog Protocol is a required configuration parameter."
+            self.logger.error(f"{validation_err_msg} {err_msg}")
             return ValidationResult(
-                success=False, message="Syslog Format is a required field."
+                success=False,
+                message=err_msg,
             )
         elif (
-            type(configuration["syslog_format"]) != str
-            or configuration["syslog_format"] not in SYSLOG_FORMATS
+            not isinstance(syslog_protocol, str)
+            or syslog_protocol not in SYSLOG_PROTOCOLS
         ):
-            self.logger.error(
-                f"{PLUGIN_NAME}[{self.name}]: Validation error occurred. Error: "
-                "Invalid Syslog Format found in the configuration parameters."
-            )
+            err_msg = "Invalid Syslog Protocol provided in configuration parameters."
+            self.logger.error(f"{validation_err_msg} {err_msg}")
             return ValidationResult(
-                success=False, message="Invalid Syslog Format provided."
+                success=False,
+                message=err_msg,
             )
-        if (
-            "syslog_protocol" not in configuration
-            or not configuration["syslog_protocol"].strip()
-        ):
-            self.logger.error(
-                f"{PLUGIN_NAME}[{self.name}]: Validation error occurred. Error: "
-                "Syslog Protocol is a required field in \
-                    the configuration parameters."
-            )
+
+        syslog_port = configuration.get("syslog_port")
+        if not syslog_port:
+            err_msg = "Syslog Port is a required configuration parameter."
+            self.logger.error(f"{validation_err_msg} {err_msg}")
             return ValidationResult(
-                success=False, message="Syslog Protocol is a required field."
+                success=False,
+                message=err_msg,
             )
-        elif (
-            type(configuration["syslog_protocol"]) != str
-            or configuration["syslog_protocol"] not in SYSLOG_PROTOCOLS
-        ):
-            self.logger.error(
-                f"{PLUGIN_NAME}[{self.name}]: Validation error occurred. Error: "
-                "Invalid Syslog Protocol found in \
-                    the configuration parameters."
-            )
+        elif not isinstance(syslog_port, int):
+            err_msg = "Invalid Syslog Port provided in configuration parameters."
+            self.logger.error(f"{validation_err_msg} {err_msg}")
             return ValidationResult(
-                success=False, message="Invalid Syslog Protocol provided."
+                success=False,
+                message=err_msg,
             )
-        if (
-            "syslog_port" not in configuration
-            or not configuration["syslog_port"]
-        ):
-            self.logger.error(
-                f"{PLUGIN_NAME}[{self.name}]: Validation error occurred. Error: "
-                "Syslog Port is a required field in \
-                    the configuration parameters."
-            )
+        elif not syslog_validator.validate_syslog_port(syslog_port):
+            err_msg = "Invalid Syslog Port provided in configuration parameters. it should be in range 0-65535."
+            self.logger.error(f"{validation_err_msg} {err_msg}")
             return ValidationResult(
-                success=False, message="Syslog Port is a required field."
+                success=False,
+                message=err_msg,
             )
-        elif not syslog_validator.validate_syslog_port(
-            configuration["syslog_port"]
-        ):
-            self.logger.error(
-                f"{PLUGIN_NAME}[{self.name}]: Validation error occurred. Error: "
-                "Invalid Syslog Port found in the configuration parameters."
-            )
-            return ValidationResult(
-                success=False, message="Invalid Syslog Port provided."
-            )
+
         mappings = self.mappings.get("jsonData", None)
         mappings = json.loads(mappings)
-        if type(mappings) != dict or not syslog_validator.validate_syslog_map(
+        if not isinstance(mappings, dict) or not syslog_validator.validate_syslog_map(
             mappings
         ):
-            self.logger.error(
-                f"{PLUGIN_NAME}[{self.name}]: Validation error occurred. Error: "
-                "Invalid Syslog attribute mapping found in \
-                    the configuration parameters."
-            )
+            err_msg = "Invalid attribute mapping provided."
+            self.logger.error(f"{validation_err_msg} {err_msg}")
             return ValidationResult(
                 success=False,
-                message="Invalid Syslog attribute mapping provided.",
+                message=err_msg,
             )
-        if configuration["syslog_protocol"].upper() == "TLS" and (
-            "syslog_certificate" not in configuration
-            or not configuration["syslog_certificate"].strip()
-        ):
-            self.logger.error(
-                f"{PLUGIN_NAME}[{self.name}]: Validation error occurred. Error: "
-                "Syslog Certificate mapping is a required field when TLS is \
-                      provided in the configuration parameters."
-            )
+
+        syslog_certificate = configuration.get("syslog_certificate", "").strip()
+        if syslog_protocol.upper() == "TLS":
+            if not syslog_certificate:
+                err_msg = (
+                    "Syslog Certificate is a required configuration parameter when "
+                    "TLS is provided in the configuration parameters."
+                )
+                self.logger.error(f"{validation_err_msg} {err_msg}")
+                return ValidationResult(
+                    success=False,
+                    message=err_msg,
+                )
+            elif not isinstance(syslog_certificate, str):
+                err_msg = (
+                    "Invalid Syslog Certificate provided in configuration parameters."
+                )
+                self.logger.error(f"{validation_err_msg} {err_msg}")
+                return ValidationResult(
+                    success=False,
+                    message=err_msg,
+                )
+
+        log_source_identifier = configuration.get("log_source_identifier", "").strip()
+        if not log_source_identifier:
+            err_msg = "Log Source Identifier is a required configuration parameter."
+            self.logger.error(f"{validation_err_msg} {err_msg}")
             return ValidationResult(
                 success=False,
-                message="Syslog Certificate mapping is a required \
-                    field when TLS is provided.",
+                message=err_msg,
             )
-        elif (
-            configuration["syslog_protocol"].upper() == "TLS"
-            and type(configuration["syslog_certificate"]) != str
-        ):
-            self.logger.error(
-                f"{PLUGIN_NAME}[{self.name}]: Validation error occurred. Error: "
-                "Invalid Syslog Certificate mapping found in \
-                    the configuration parameters."
+        elif not isinstance(log_source_identifier, str):
+            err_msg = (
+                "Invalid Log Source Identifier provided in configuration parameters."
             )
+            self.logger.error(f"{validation_err_msg} {err_msg}")
             return ValidationResult(
                 success=False,
-                message="Invalid Syslog Certificate mapping provided.",
-            )
-        if (
-            "log_source_identifier" not in configuration
-            or not configuration["log_source_identifier"].strip()
-        ):
-            self.logger.error(
-                f"{PLUGIN_NAME}[{self.name}]: Validation error occurred. Error: "
-                "Log Source Identifier is a required field in \
-                    the configuration parameters."
-            )
-            return ValidationResult(
-                success=False,
-                message="Log Source Identifier is a required field.",
-            )
-        elif (
-            type(configuration["log_source_identifier"]) != str
-            or " " in configuration["log_source_identifier"].strip()
-        ):
-            self.logger.error(
-                f"{PLUGIN_NAME}[{self.name}]: Validation error occurred. Error: "
-                "Invalid Log Source Identifier found in \
-                    the configuration parameters."
-            )
-            return ValidationResult(
-                success=False,
-                message="Invalid Log Source Identifier provided.",
+                message=err_msg,
             )
         # Validate Server connection.
         try:
             self.test_server_connectivity(configuration)
-        except Exception:
-            self.logger.error(
-                f"{PLUGIN_NAME}[{self.name}]: Validation error occurred. Error: "
-                "Connection to SIEM platform is not established."
-            )
+        except SyslogPluginError as err:
             return ValidationResult(
                 success=False,
-                message="Error occurred while establishing connection \
-                with Syslog Server. Make sure you have provided correct \
-                Syslog Server, Port and Syslog Certificate(if required).",
+                message=str(err),
             )
+        except Exception:
+            err_msg = (
+                "Error occurred while establishing connection with Syslog Server. "
+                "Make sure you have provided correct Syslog Server, Port and "
+                "Syslog Certificate(if required)."
+            )
+            self.logger.error(f"{validation_err_msg} {err_msg}")
+            return ValidationResult(
+                success=False,
+                message=err_msg,
+            )
+
         return ValidationResult(success=True, message="Validation successful.")
 
     def chunk_size(self):