+++ syslog/utils/syslog_cef_generator.py
@@ -28,21 +28,18 @@ SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
 CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY,
 OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
 OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-"""
 
-"""Syslog Plugin."""
-
+Syslog Plugin."""
 
 import collections
-import socket
 import time
-
+import traceback
 from .syslog_constants import (
     SEVERITY_MAP,
     SEVERITY_UNKNOWN,
     AUDIT_SEVERITY_MAP,
-    PLUGIN_NAME,
 )
+from .syslog_exceptions import SyslogPluginError
 from netskope.integrations.cls.utils.sanitizer import *
 from netskope.integrations.cls.utils.converter import *
 
@@ -50,15 +47,13 @@ from netskope.integrations.cls.utils.converter import *
 class CEFGenerator(object):
     """CEF Generator class."""
 
-    def __init__(self, mapping, delimiter, cef_version, logger, name):
+    def __init__(self, mapping, delimiter, cef_version, logger, log_prefix):
         """Init method."""
         self.logger = logger
-        self.name = name
+        self.log_prefix = log_prefix
         self.cef_version = cef_version  # Version of CEF being used
         self.mapping = mapping  # Mapping file content
-        self.extension = collections.namedtuple(
-            "Extension", ("key_name", "sanitizer")
-        )
+        self.extension = collections.namedtuple("Extension", ("key_name", "sanitizer"))
         self.extension_converter = collections.namedtuple(
             "Extension", ("key_name", "converter")
         )
@@ -75,10 +70,12 @@ class CEFGenerator(object):
         self.delimiter = delimiter
 
     def _type_converter(self):
-        """To Parse the CEF transformation mapping and creates the dict for data type converters.
+        """To Parse the CEF transformation mapping and creates
+        the dict for data type converters.
 
         Returns:
-            Dict object having details of all the available CEF fields and its type converters
+            Dict object having details of all the available CEF fields and
+            its type converters
         """
         converters = type_converter()
 
@@ -96,23 +93,30 @@ class CEFGenerator(object):
                             field_converters[field] = self.extension_converter(
                                 key_name=field,
                                 converter=converters[
-                                    field_mapping.get(
-                                        "transformation", "String"
-                                    )
+                                    field_mapping.get("transformation", "String")
                                 ],
                             )
             return field_converters
         except Exception as err:
-            self.logger.error(
-                f"{PLUGIN_NAME}[{self.name}]: Error occurred while parsing CEF transformation field. Error: {err}"
+            error_msg = (
+                "Error occurred while parsing CEF "
+                "transformation field."
             )
-            raise
+            self.logger.error(
+                message=(
+                    f"{self.log_prefix}: {error_msg} Error: {err}"
+                ),
+                details=str(traceback.format_exc()),
+            )
+            raise SyslogPluginError(error_msg)
 
     def _valid_extensions(self):
-        """To Parse the given transformation mapping and creates the dict for each provided values with its sanitizers.
+        """To Parse the given transformation mapping and creates
+        the dict for each provided values with its sanitizers.
 
         Returns:
-            Dict object having details of all the available CEF fields and its sanitizers
+            Dict object having details of all the available CEF fields and
+            its sanitizers
         """
         sanitizers = get_sanitizers()
 
@@ -130,21 +134,26 @@ class CEFGenerator(object):
                             field_sanitizers[field] = self.extension(
                                 key_name=field,
                                 sanitizer=sanitizers[
-                                    field_mapping.get(
-                                        "transformation", "String"
-                                    )
+                                    field_mapping.get("transformation", "String")
                                 ],
                             )
             return field_sanitizers
         except Exception as err:
-            self.logger.error(
-                f"{PLUGIN_NAME}[{self.name}]: Error occurred while parsing CEF transformation field. \
-                    Error: {err}"
+            error_msg = (
+                "Error occurred while parsing CEF "
+                "transformation field."
             )
-            raise
+            self.logger.error(
+                message=(
+                    f"{self.log_prefix}: {error_msg} Error: {err}"
+                ),
+                details=str(traceback.format_exc()),
+            )
+            raise SyslogPluginError(error_msg)
 
     def get_header_value(self, header, headers):
-        """To Fetch sanitized value of header from given configured headers dict.
+        """To Fetch sanitized value of header from given
+        configured headers dict.
 
         Args:
             header: The header for which sanitized value is to be fetched
@@ -157,22 +166,22 @@ class CEFGenerator(object):
             return self._severity_sanitizer(headers[header], header)
         return self._prefix_field_str_sanitizer(headers[header], header)
 
-    def log_invalid_header(
-        self, possible_headers, headers, data_type, subtype
-    ):
+    def log_invalid_header(self, possible_headers, headers, data_type, subtype):
         """Issues log in case of invalid header found in mappings.
 
         Args:
             possible_headers: Possible CEF headers
             headers: Configured headers
             data_type: Data type for which CEF event is being generated
-            subtype: Subtype of data type for which CEF event is being generated
+            subtype: Subtype of data type for which
+            CEF event is being generated
         """
         for configured_header in list(headers.keys()):
             if configured_header not in possible_headers:
-                self.logger.error(
-                    f'{PLUGIN_NAME}[{self.configuration["name"]}]([{data_type}][{subtype}]): Found invalid header configured in syslog mapping file: "{configured_header}". Header '
-                    "field will be ignored."
+                self.logger.warn(
+                    f"{self.log_prefix}: [{data_type}][{subtype}]- "
+                    "Found invalid header configured in syslog mapping file: "
+                    f"{configured_header}. Header field will be ignored."
                 )
 
     def webtx_timestamp(self, raw_data):
@@ -202,42 +211,52 @@ class CEFGenerator(object):
         """
         extension_strs = {}
         for name, value in extensions.items():
-            # First convert the incoming value from Netskope to appropriate data type
+            # First convert the incoming value from
+            # Netskope to appropriate data type
             try:
                 value = self.extension_converters[name].converter(value, name)
             except KeyError:
-                self.logger.error(
-                    f'{PLUGIN_NAME}[{self.configuration["name"]}]([{data_type}][{subtype}]): An error occurred while generating CEF data for field: "{name}". Could not '
-                    'find the field in the "valid_extensions". Field will be ignored'
+                self.logger.warn(
+                    "{}: [{}][{}] An error occurred while generating CEF "
+                    'data for field: "{}". Could not find the field in '
+                    'the "valid_extensions". Field will be ignored.'.format(
+                        self.log_prefix, data_type, subtype, name
+                    )
                 )
                 continue
             except Exception as err:
-                self.logger.error(
-                    f'{PLUGIN_NAME}[{self.configuration["name"]}]([{data_type}][{subtype}]): An error occurred while generating CEF data for field: "{name}". Error: {err}. '
-                    "Field will be ignored"
+                self.logger.warn(
+                    "{}: [{}][{}] An error occurred while generating "
+                    'CEF data for field: "{}". Error: {}. '
+                    "Field will be ignored.".format(
+                        self.log_prefix, data_type, subtype, name, err
+                    )
                 )
                 continue
 
-            # Validate and sanitise (if required) the incoming value from Netskope before mapping it CEF
+            # Validate and sanitise (if required) the incoming value
+            # from Netskope before mapping it CEF
             try:
-                sanitized_value = self.valid_extensions[name].sanitizer(
-                    value, name
-                )
+                sanitized_value = self.valid_extensions[name].sanitizer(value, name)
                 if isinstance(sanitized_value, str):
                     sanitized_value = self._equals_escaper(sanitized_value)
 
-                extension_strs[
-                    self.valid_extensions[name].key_name
-                ] = sanitized_value
+                extension_strs[self.valid_extensions[name].key_name] = sanitized_value
             except KeyError:
-                self.logger.error(
-                    f'{PLUGIN_NAME}[{self.configuration["name"]}]([{data_type}][{subtype}]): An error occurred while generating CEF data for field: "{name}". Could not '
-                    'find the field in the "valid_extensions". Field will be ignored'
+                self.logger.warn(
+                    "{}: [{}][{}] An error occurred while generating "
+                    'CEF data for field: "{}". Could not find the field in '
+                    'the "valid_extensions". Field will be ignored.'.format(
+                        self.log_prefix, data_type, subtype, name
+                    )
                 )
             except Exception as err:
-                self.logger.error(
-                    f'{PLUGIN_NAME}[{self.configuration["name"]}]([{data_type}][{subtype}]): An error occurred while generating CEF data for field: "{name}". Error: {err}. '
-                    "Field will be ignored"
+                self.logger.warn(
+                    "{}: [{}][{}] An error occurred while generating CEF "
+                    'data for field: "{}". Error: {}. '
+                    "Field will be ignored.".format(
+                        self.log_prefix, data_type, subtype, name, err
+                    )
                 )
 
         possible_headers = [
@@ -276,14 +295,15 @@ class CEFGenerator(object):
                                 str(headers[header]).lower(), SEVERITY_UNKNOWN
                             )
 
-                    cef_components.append(
-                        self.get_header_value(header, headers)
-                    )
+                    cef_components.append(self.get_header_value(header, headers))
 
                 except Exception as err:
-                    self.logger.error(
-                        f'{PLUGIN_NAME}[{self.configuration["name"]}]([{data_type}][{subtype}]): An error occurred while generating CEF data for header field: "{header}". Error: {err}. '
-                        "Field will be ignored"
+                    self.logger.warn(
+                        "{}: [{}][{}] An error occurred while generating "
+                        'CEF data for header field: "{}". Error: {}. '
+                        "Field will be ignored.".format(
+                            self.log_prefix, data_type, subtype, header, err
+                        )
                     )
 
         if data_type == "webtx":