+++ syslog/utils/syslog_exceptions.py
@@ -28,9 +28,14 @@ SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
 CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY,
 OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
 OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-"""
 
-"""Exception class for syslog plugin."""
+Exception class for syslog plugin."""
+
+
+class SyslogPluginError(Exception):
+    """CEF Format Error class."""
+
+    pass
 
 
 class FormatCEFError(Exception):
@@ -46,7 +51,8 @@ class Error(Exception):
 
 
 class ValidationError(Error):
-    """Exception raised for validation failures for the parameters of docker-compose file.
+    """Exception raised for validation failures for
+    the parameters of docker-compose file.
 
     :attribute: expression -- input expression in which the error occurred
     :attribute: message -- explanation of the error
@@ -71,7 +77,8 @@ class CEFValueError(FormatCEFError, ValueError):
 
 
 class CEFTypeError(FormatCEFError, TypeError):
-    """Exception raised for data type mismatch between mapped value and CEF fields.
+    """Exception raised for data type mismatch between
+    mapped value and CEF fields.
 
     Attributes:
         message -- explanation of the error